sendMessage() works basically, initData() noch fixen
