+++ src/main/java/fhtw/user/User.java
@@ -46,17 +46,6 @@ public class User implements Serializable {
 
 
 
-/*public Set<PrivateChat> getUserPrivateChats() {
-        Set<PrivateChat> privateChats = new HashSet<>();
-        for (Integer chatId : userPrivateChatIDs) {
-            PrivateChat chat = DatabaseHandler.getPrivateChats().get(chatId);
-            if (chat != null) {
-                privateChats.add(chat);
-            }
-        }
-        return privateChats;
-    }*/
-
     public User findUser(String username) {
         if (username == null || username.isEmpty()) {
             return null;
@@ -66,25 +55,6 @@ public class User implements Serializable {
             return searchedUser;
         }
         return null;
-    }
-
-    public static boolean addUser(String sender, String receiver) {
-        User senderUser = DatabaseHandler.getRegisteredUsers().get(sender);
-        User receiverUser = DatabaseHandler.getRegisteredUsers().get(receiver);
-        System.out.println(sender);
-        System.out.println(receiver);
-        if (senderUser == null || receiverUser == null) {
-            System.out.println("SENDER "+senderUser);
-            System.out.println("Receiver "+receiverUser);
-            return false;
-        }
-        PrivateChat chat = new PrivateChat(sender, receiver);
-        DatabaseHandler.getPrivateChats().put(chat.getChatId(), chat); //adding privatchatroom to UserService HashMap
-        System.out.print("HashMAPS ");
-        System.out.println(DatabaseHandler.getPrivateChats());
-        senderUser.getPrivateChats().add(chat);
-        receiverUser.getPrivateChats().add(chat);
-        return true;
     }
